+++ tests/cli/test_benchmark_plot.py
@@ -24,14 +24,8 @@ def test_benchmark_computational_plot():
         )
 
         assert result.returncode == 0
-        assert (
-            "process-P1-.317a506603d7cb7f079fcc6a38cdf99e3955e1729540d38b9b0f36bd7c16d2a3-after_data"
-            in result.stdout
-        )
-        assert (
-            "methods-M2-.3297cc0b9f48521ab602a4a90143602c416f1f5029c70182a62e6092166d3bc9-after_data"
-            in result.stdout
-        )
+        assert "process-P1-.317a5066-after_data" in result.stdout
+        assert "methods-M2-.3297cc0b-after_data" in result.stdout
         assert "metrics-m3-default-after_methods" in result.stdout