+++ CHANGELOG.md
@@ -5,13 +5,24 @@ This project adheres to [Semantic Versioning](https://semver.org/) and [Conventi
 
 ## [0.4.0](main) UNRELEASED
 
+- feat: add create benchmark & create module command (#171)
+- feat: add ob status (#184)
+- feat: give contextual information about parsing errors (#188)
+- docs: add self-documenting spec template
 - refactor!: remove dependency on omni-schema and networkx
 - refactor: cleanup internal module dependencies, using a central benchmark model
 - refactor: use strict type checking (#70)
-- docs: add self-documenting spec template
 - tests: use pixi on the CI
-- feat: add create benchmark & create module command (#171)
+- bug: do not fail if numeric values given in commit, params or version (#187)
 
+## 0.3.2 (Nov 13th 2025)
+
+- hotfix: do not hardcode /tmp in run_module (#173)
+
+## 0.3.1 (Oct 27th 2025)
+
+- hotfix: remove the need to define dummy software environments for unused backends
+- hotfix: validate yaml against ambiguous output patterns
 
 ## [0.3.0](https://github.com/omnibenchmark/omnibenchmark/releases/tag/v0.3.0) (Aug 15th 2025)